Replace 27_Conv3d_HardSwish_ReLU_Softmax_Mean with GroupNorm variant
