+++ KernelBench/level2/27_Conv3d_HardSwish_GroupNorm_Mean.py
@@ -1,0 +1,36 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+class Model(nn.Module):
+    """
+    Model that performs:
+    1. Conv3D
+    2. HardSwish activation
+    3. GroupNorm  
+    4. Mean pooling across spatial dimensions
+    """
+    def __init__(self, in_channels, out_channels, kernel_size, num_groups=4, bias=True):
+        super(Model, self).__init__()
+        self.conv = nn.Conv3d(in_channels, out_channels, kernel_size, bias=bias)
+        self.group_norm = nn.GroupNorm(num_groups, out_channels)
+
+    def forward(self, x):
+        x = self.conv(x)                             # (B, C, D, H, W)
+        x = F.hardswish(x)                           # Nonlinear activation
+        x = self.group_norm(x)                       # Normalization over channels
+        x = torch.mean(x, dim=[2, 3, 4])             # Mean over spatial dims → (B, C)
+        return x
+
+# === Test config ===
+batch_size = 128
+in_channels = 3
+out_channels = 16
+depth, height, width = 16, 32, 32
+kernel_size = 3
+
+def get_inputs():
+    return [torch.randn(batch_size, in_channels, depth, height, width)]
+
+def get_init_inputs():
+    return [in_channels, out_channels, kernel_size]
